+++ src/file/memfd.h
@@ -2,11 +2,17 @@
 #include "common/bpf_helpers.h"
 #include "common/types.h"
 #include "maps.h"
-#include "modify.h"
 #include "push_file_message.h"
 #include "vmlinux.h"
 
 #include "common/definitions.h"
+
+struct syscalls_enter_memfd_create_args {
+   __u64 unused;
+   long __syscall_nr;
+   const char *uname;
+   unsigned long flags;
+};
 
 static const char MEMFD_PREFIX[] = "memfd:";
