+++ src/file/open.h
@@ -2,13 +2,6 @@
 
 #include "common/definitions.h"
 
-struct syscalls_enter_memfd_create_args {
-   __u64 unused;
-   long __syscall_nr;
-   const char *uname;
-   unsigned long flags;
-};
-
 struct syscalls_enter_open_args {
     __u64 unused;
     long __syscall_nr;
